+++ Arcadia/Classes/ArcadiaFileManager.swift
@@ -155,6 +155,30 @@ enum ArcadiaCloudSyncStatus {
         }
     }
     
+    func importSaveFile(for gameURL: URL, saveURL: URL, gameType: ArcadiaGameType, needScope: Bool = true) {
+        if needScope {
+            if gameURL.startAccessingSecurityScopedResource()  {
+                defer {
+                    gameURL.stopAccessingSecurityScopedResource()
+                }
+                
+                let localSaveURL = getSaveURL(gameURL: gameURL, gameType: gameType)
+                
+                do {
+                    let saveFile = try Data(contentsOf: saveURL)
+                    try saveFile.write(to: localSaveURL, options: .atomic)
+                    if let iCloudSyncEnabled = UserDefaults.standard.object(forKey: "iCloudSyncEnabled") as? Bool {
+                        if iCloudSyncEnabled {
+                            createCloudCopy(of: localSaveURL)
+                        }
+                    }
+                } catch {
+                    print("couldn't save file \(error)")
+                }
+            }
+        }
+    }
+    
     func saveGame(gameURL: URL, gameType: ArcadiaGameType, needScope: Bool = true) {
         if needScope {
             if gameURL.startAccessingSecurityScopedResource()  {
@@ -222,8 +246,10 @@ enum ArcadiaCloudSyncStatus {
                     }
                 }
                 //To update the game list
-                if needScope {
-                    getGamesURL(gameSystem: gameType)
+                if let currentGameSystem = ArcadiaNavigationState.shared.currentGameSystem {
+                    if currentGameSystem == gameType {
+                        getGamesURL(gameSystem: gameType)
+                    }
                 }
             } catch {
                 print("couldn't save file \(error)")
@@ -244,7 +270,7 @@ enum ArcadiaCloudSyncStatus {
     }
     
     func getSaveURL(gameURL: URL, gameType: ArcadiaGameType) -> URL {
-        return self.savesDirectory.appendingPathComponent(gameType.rawValue).appendingPathComponent(gameURL.deletingPathExtension().lastPathComponent).appendingPathExtension("srm")
+        return self.savesDirectory.appendingPathComponent(gameType.rawValue).appendingPathComponent(gameURL.deletingPathExtension().lastPathComponent).appendingPathExtension(gameType.supportedSaveFiles[.memorySaveRam] ?? "srm")
     }