+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.4
+
+## Features
+
+- Added the possibility to open game files with Arcadia from the file's share sheet
+- Added the possibility to share games and export save files
+
+## Improvements
+
+- Improved the aspect of the empty library view
+- Added the possibility to refresh the empty library view
+
+## Bug fixes
+
+
 # 1.3
 
 ## Features